requirements.txt
streamlit
pandas
numpy
scikit-learn
chromadb
sentence-transformers
openai>=1.0.0
python-dotenv
tqdm
uvicorn
fastapi